requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-admin-bootstrapped==2.5.6
django-bootstrap3==6.2.2
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
whitenoise==2.0.4
djangorestframework==3.3.2
Markdown==2.6.5
django-filter==0.12.0
requests==2.9.1
simplejson==3.8.2

